Vagrant: Use static idstart value to ensure consistency between replicas
By default the idstat value is randomly generated if not defined. This means
that the value will be different between ipa-server-1 and ipa-server-2, because
the installer will get different values on each. This caused /home/admin on the
ipa-server-2 to have wrong UID and GID, which caused puppet run failures and
left the node into inconsistent state.