Migrate package names to params.pp
The package name depends on operating system and there should not usually be any
need to define them manually. If this module is missing correct package names
for a platform then support should be added to params.pp instead of working
around the problem by manually defining package names in Hiera or equivalent.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>